refactor(Tree): solve circular import (#38421)
